Fixed 64-bit related compiler warnings

+++ Classes/UVComment.m
@@ -25,9 +25,9 @@
 @synthesize createdAt;
 
 + (id)getWithSuggestion:(UVSuggestion *)suggestion page:(NSInteger)page delegate:(id)delegate {
-    NSString *path = [self apiPath:[NSString stringWithFormat:@"/forums/%d/suggestions/%d/comments.json",
-                                    suggestion.forumId,
-                                    suggestion.suggestionId]];
+    NSString *path = [self apiPath:[NSString stringWithFormat:@"/forums/%ld/suggestions/%ld/comments.json",
+                                    (long)suggestion.forumId,
+                                    (long)suggestion.suggestionId]];
     NSDictionary *params = [NSDictionary dictionaryWithObjectsAndKeys:
                             [[NSNumber numberWithInt:page] stringValue],
                             @"page",
@@ -40,9 +40,9 @@
 }
 
 + (id)createWithSuggestion:(UVSuggestion *)suggestion text:(NSString *)text delegate:(id)delegate {
-    NSString *path = [self apiPath:[NSString stringWithFormat:@"/forums/%d/suggestions/%d/comments.json",
-                                    suggestion.forumId,
-                                    suggestion.suggestionId]];
+    NSString *path = [self apiPath:[NSString stringWithFormat:@"/forums/%ld/suggestions/%ld/comments.json",
+                                    (long)suggestion.forumId,
+                                    (long)suggestion.suggestionId]];
     NSDictionary *params = [NSDictionary dictionaryWithObjectsAndKeys:
                             text, @"comment[text]",
                             nil];
